% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildcatdown.R
\docType{package}
\name{wildcatdown}
\alias{wildcatdown}
\alias{wildcatdown-package}
\title{wildcatdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
wildcatdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown
}
